+++ src/test/groovy/ru/vyarus/gradle/plugin/github/configurer/PomConfigurerKitTest.groovy
@@ -54,9 +54,9 @@ class PomConfigurerKitTest extends AbstractKitTest {
         then: "scm section applied"
         def projectId = "test/$testProjectDir.root.name"
         def scm = pom.scm
-        scm.url.text() == "https://github.com/${projectId}.git"
-        scm.connection.text() == "scm:git:git://github.com/${projectId}.git"
-        scm.developerConnection.text() == "scm:git:git://github.com/${projectId}.git"
+        scm.url.text() == "https://github.com/${projectId}"
+        scm.connection.text() == "scm:git:git://github.com/${projectId}"
+        scm.developerConnection.text() == "scm:git:git://github.com/${projectId}"
 
         then: "license section applied"
         def license = pom.licenses.license
@@ -123,8 +123,8 @@ class PomConfigurerKitTest extends AbstractKitTest {
         def projectId = "test/$testProjectDir.root.name"
         def scm = pom.scm
         scm.url.text() == "http://other.url"
-        scm.connection.text() == "scm:git:git://github.com/${projectId}.git"
-        scm.developerConnection.text() == "scm:git:git://github.com/${projectId}.git"
+        scm.connection.text() == "scm:git:git://github.com/${projectId}"
+        scm.developerConnection.text() == "scm:git:git://github.com/${projectId}"
 
         then: "license section applied"
         def license = pom.licenses.license